🧀 refactor: structures
